pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["artem <ki4igin1992@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyserial = "^3.5"
crcmod = "^1.7"


[tool.poetry.group.dev.dependencies]
ipython = "^8.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
